Move Autocompletable and Optionality to another package
